Workaround SDL 2.24.0 issue with low sample rate
An audio issue was introduced when upgrading from SDL 2.0.22 to 2.24.0.
If we open an audio device in SDL with a low sample rate, this causes
all audio produced on the computer to become distorted (including audio
produced by apps other than Moonlight).

The workaround in this commit is simply to set a minimum sample rate of
480, at which point the problem disappears. The magic value 480 was
determined empirically, and slightly higher or slightly lower values
would probably work equally well. The default value before this
workaround was 240.