Add GetCurrentMbps() method to BandwidthTracker
Fixes compilation error when DISPLAY_BITRATE is enabled. The code in
ffmpeg.cpp:873 calls GetCurrentMbps() but this method was missing from
the BandwidthTracker class.

Implementation returns the bitrate of the most recently completed bucket,
providing a current value that is more recent than the average but more
stable than an in-progress bucket.